moved get_block_events to events.rs
